requirements.txt
python-telegram-bot
requests
ta
tenacity
ccxt
shap
matplotlib
pytest
flake8
lightgbm
scikit-optimize
xgboost
imblearn
pytest-responses
apscheduler
python-dotenv
threadpoolctl-3.5.0
aiodns              3.2.0
aiohappyeyeballs    2.4.4
aiohttp             3.10.11
aiosignal           1.3.2
anyio               4.8.0
APScheduler         3.11.0
attrs               24.3.0
ccxt                4.4.47
certifi             2024.12.14
cffi                1.17.1
charset-normalizer  3.4.1
colorama            0.4.6
contourpy           1.3.1
cryptography        44.0.0
cycler              0.12.1
fonttools           4.55.3
frozenlist          1.5.0
h11                 0.14.0
httpcore            1.0.7
httpx               0.28.1
idna                3.10
iniconfig           2.0.0
joblib              1.4.2
kiwisolver          1.4.8
matplotlib          3.10.0
multidict           6.1.0
numpy               2.2.1
packaging           24.2
pandas              2.2.3
pillow              11.1.0
pip                 24.3.1
pluggy              1.5.0
propcache           0.2.1
pycares             4.5.0
pycparser           2.22
pyparsing           3.2.1
pytest              8.3.4
pytest-asyncio      0.25.2
python-dateutil     2.9.0.post0
python-dotenv       1.0.1
python-telegram-bot 21.10
pytz                2024.2
requests            2.32.3
scikit-learn        1.6.1
scipy               1.15.1
setuptools          75.8.0
six                 1.17.0
sniffio             1.3.1
ta                  0.11.0
TA-Lib              0.4.32
threadpoolctl       3.5.0
typing_extensions   4.12.2
tzdata              2024.2
tzlocal             5.2
urllib3             2.3.0
websocket-client    1.8.0
yarl                1.18.3

